task_1/code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_py-xgboost-mutex=2.0=cpu_0
anyio=3.5.0=py39hecd8cb5_0
appdirs=1.4.4=pyhd3eb1b0_0
appnope=0.1.2=py39hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39hca72f7f_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py39hecd8cb5_0
babel=2.11.0=py39hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py39h67323c0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py39h9ed2024_1003
ca-certificates=2023.05.30=hecd8cb5_0
cairo=1.16.0=h3ce6f7e_4
catboost=1.2=py39hecd8cb5_0
certifi=2023.5.7=py39hecd8cb5_0
cffi=1.15.1=py39h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.2=py39hecd8cb5_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py39hecd8cb5_0
exceptiongroup=1.0.4=py39hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=he9d5cce_0
flit-core=3.6.0=pyhd3eb1b0_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=hedf32ac_1
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.39.2=pypi_0
freetype=2.12.1=hd8bbffd_0
fribidi=1.0.10=haf1e3a3_0
gdk-pixbuf=2.42.10=h6c40b1e_0
gettext=0.21.0=h7535e17_0
gif=23.3.0=pypi_0
giflib=5.2.1=h6c40b1e_3
glib=2.69.1=hfff2838_2
graphite2=1.3.14=he9d5cce_1
graphviz=2.50.0=h196fa6a_0
gst-plugins-base=1.14.1=hcec6c5f_1
gstreamer=1.14.1=h6c40b1e_1
gts=0.7.6=h6759243_3
harfbuzz=4.3.0=hffc734d_1
icu=58.2=h0a44026_3
idna=3.4=py39hecd8cb5_0
importlib-metadata=6.0.0=py39hecd8cb5_0
importlib-resources=5.12.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.19.2=py39h01d92e1_0
ipython=8.10.0=py39hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py39hecd8cb5_0
jedi=0.18.1=py39hecd8cb5_1
jinja2=3.1.2=py39hecd8cb5_0
joblib=1.1.1=py39hecd8cb5_0
jpeg=9e=h6c40b1e_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py39hecd8cb5_0
jupyter=1.0.0=py39hecd8cb5_8
jupyter_client=7.4.9=py39hecd8cb5_0
jupyter_console=6.6.2=py39hecd8cb5_0
jupyter_core=5.2.0=py39hecd8cb5_0
jupyter_server=1.23.4=py39hecd8cb5_0
jupyterlab=3.5.3=py39hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.19.0=py39hecd8cb5_0
jupyterlab_widgets=3.0.5=py39hecd8cb5_0
kaleido=0.2.1=pypi_0
kiwisolver=1.4.4=py39hcec6c5f_0
krb5=1.19.4=hdba6334_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=12.0.0=default_hbc2896b_2
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libedit=3.1.20221030=h6c40b1e_0
libffi=3.4.2=hecd8cb5_6
libgd=2.3.3=hcec6c5f_2
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libllvm12=12.0.0=h9b2ccf5_3
libpng=1.6.39=h6c40b1e_0
libpq=12.9=h1c9f633_3
librsvg=2.54.4=h2f8da11_2
libsodium=1.0.18=h1de35cc_0
libtiff=4.5.0=hcec6c5f_2
libtool=2.4.6=hcec6c5f_1009
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxgboost=1.7.3=hcec6c5f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
llvm-openmp=14.0.6=h0dcd299_0
lxml=4.9.1=py39h65b224f_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.1=py39hca72f7f_0
matplotlib-base=3.7.1=py39hda11e5a_1
matplotlib-inline=0.1.6=py39hecd8cb5_0
mistune=0.8.4=py39h9ed2024_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
nb_conda=2.2.1=py39hecd8cb5_1
nb_conda_kernels=2.3.1=py39hecd8cb5_0
nbclassic=0.5.2=py39hecd8cb5_0
nbclient=0.5.13=py39hecd8cb5_0
nbconvert=6.5.4=py39hecd8cb5_0
nbformat=5.7.0=py39hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py39hecd8cb5_0
notebook=6.5.2=py39hecd8cb5_0
notebook-shim=0.2.2=py39hecd8cb5_0
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numexpr=2.8.4=py39he696674_0
numpy=1.23.5=py39he696674_0
numpy-base=1.23.5=py39h9cd3388_0
openssl=1.1.1t=hca72f7f_0
packaging=23.0=py39hecd8cb5_0
pandas=1.5.3=py39h07fba90_0
pandocfilters=1.5.0=pyhd3eb1b0_0
pango=1.50.7=h80fe9ab_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.3=py39hecd8cb5_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=pypi_0
pip=23.0.1=py39hecd8cb5_0
pixman=0.40.0=h9ed2024_1
platformdirs=2.5.2=py39hecd8cb5_0
plotly=5.9.0=py39hecd8cb5_0
pluggy=1.0.0=py39hecd8cb5_1
ply=3.11=py39hecd8cb5_0
pooch=1.4.0=pyhd3eb1b0_0
prometheus_client=0.14.1=py39hecd8cb5_0
prompt-toolkit=3.0.36=py39hecd8cb5_0
prompt_toolkit=3.0.36=hd3eb1b0_0
psutil=5.9.0=py39hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-xgboost=1.7.3=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py39hecd8cb5_0
pyparsing=3.0.9=py39hecd8cb5_0
pyqt=5.15.7=py39he9d5cce_0
pyqt5-sip=12.11.0=py39he9d5cce_0
pyrsistent=0.18.0=py39hca72f7f_0
pysocks=1.7.1=py39hecd8cb5_0
pytest=7.3.1=py39hecd8cb5_0
python=3.9.16=h218abb5_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39hecd8cb5_0
python-graphviz=0.20.1=py39hecd8cb5_0
pytz=2022.7=py39hecd8cb5_0
pyzmq=23.2.0=py39he9d5cce_0
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_4
qtconsole=5.4.0=py39hecd8cb5_0
qtpy=2.2.0=py39hecd8cb5_0
qtwebkit=5.212=h24dc246_4
readline=8.2=hca72f7f_0
requests=2.28.1=py39hecd8cb5_1
scikit-learn=1.2.1=py39hcec6c5f_0
scipy=1.10.0=py39h91c6ef4_1
seaborn=0.12.2=py39hecd8cb5_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.6.3=py39hecd8cb5_0
sip=6.6.2=py39he9d5cce_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39hecd8cb5_1
soupsieve=2.3.2.post1=py39hecd8cb5_0
sqlite=3.41.1=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.5=py39hacda100_1
tenacity=8.0.1=py39hecd8cb5_1
terminado=0.17.1=py39hecd8cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py39hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py39hecd8cb5_0
tornado=6.2=py39hca72f7f_0
traitlets=5.7.1=py39hecd8cb5_0
typing-extensions=4.4.0=py39hecd8cb5_0
typing_extensions=4.4.0=py39hecd8cb5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py39hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39hecd8cb5_1
websocket-client=0.58.0=py39hecd8cb5_4
wheel=0.38.4=py39hecd8cb5_0
widgetsnbextension=4.0.5=py39hecd8cb5_0
xgboost=1.7.3=py39hecd8cb5_0
xz=5.2.10=h6c40b1e_1
zeromq=4.3.4=h23ab428_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
